Hatfield Patch 2
Updated the updatePrice method and moved it to the AbstractPizza class
so any pizza can be allowed to be updated to what it would currently
cost based on toppings and type. Does not currently consider other
factors but can be added later if needed.

Added an orderIDcounter to the PizzaCookingFactory class to give each
created object a unique incremental ID

Added attributes to pizza order and constructors

Added printListOfToppingsByPizzaOrderID method

Added getPIzzaByOrderID method

Added addPizzaToCart method

Added printPizzaOrderCart method